+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.0.6] - 2023-05-07
 
 ### Added
 
@@ -90,6 +90,7 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - 🎉 Initial release.
 
+[0.0.6]: https://github.com/sunsided/pddl-rs/releases/tag/0.0.6
 [0.0.5]: https://github.com/sunsided/pddl-rs/releases/tag/0.0.5
 [0.0.4]: https://github.com/sunsided/pddl-rs/releases/tag/0.0.4
 [0.0.3]: https://github.com/sunsided/pddl-rs/releases/tag/0.0.3